Allow skipping splits and add gdiag fitting for CS2 trees
The gdiag fitting represents an alternative means for specifying
g and the diagonal elements of Q. After setting all the off-diagonal
elements of Q, these two can be found by solving a 2x2 equation in
each coordinate. The main advantage is that this allows one to use
the splits(box) iterator directly.

Adding "skip" functionality to the fits is directed towards computing
a residual; skipping means that the residual will be less affected by
"overfitted" splits (where the model agrees with the data point by
construction). It's possible that one could exclude these more robustly,
but this is a good start.