Add lower-bound fitting model
This takes a collection of boxes and ensures that the quadratic model
is bounded below by the data points. This might be viewed as a "conservative fit"
of the function values.

This uses a linear-programming solver. Th package is agnostic about the
particular solver used, as the optimizer gets passed as an argument.